Moves Placeholder project to a folder of it's own
